% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_flipscores.R
\name{as.jfs}
\alias{as.jfs}
\title{Coerce fs_contrasts to jfs}
\usage{
as.jfs(x, ...)
}
\arguments{
\item{x}{an object of class \code{fs_contrasts}}

\item{...}{further arguments (currently unused)}
}
\value{
an object of class \code{jfs}
}
\description{
Coerces an \code{fs_contrasts} object to class \code{jfs}.
}
