% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_flipscores_methods.R
\name{c.jfs}
\alias{c.jfs}
\title{Combine jfs objects}
\usage{
\method{c}{jfs}(...)
}
\arguments{
\item{...}{Objects to combine. Supported objects are \code{jfs},
\code{flipscores}, \code{glm} and \code{fs_contrasts}.}
}
\value{
A \code{jfs} object.
}
\description{
Combine jfs objects
}
