% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcor.R
\name{compute_gcor}
\alias{compute_gcor}
\title{compute_gcor}
\usage{
compute_gcor(model0, X, ...)
}
\arguments{
\item{model0}{a \code{glm} object with the model under the null hypothesis (i.e. the covariates, the nuisance parameters).}

\item{...}{other arguments.}

\item{model1}{a \code{glm} or a \code{matrix} (or \code{vector}). If it is a \code{glm} object, it has the model under the alternative hypothesis. The variables in \code{model1} are the same variables in \code{model0} plus one or more variables to be tested.  Alternatively, if
\code{model1} is a \code{matrix}, it contains the tested variables column-wise.}

\item{score_type}{The type of score that is computed. It is "orthogonalized", "effective" or "basic".
"effective" and "orthogonalized" take into account the nuisance estimation.}
}
\description{
Same usage as \code{anova.glm}.
The parameter \code{id}  is used too,
if present in \code{model0} (with priority) or in \code{model1}.
}
\examples{
library(rsq)
set.seed(1)
Z=rnorm(20)
X=Z+rnorm(20)
Y=rpois(n=20,lambda=exp(Z+.5*X))
model0=glm(Y~Z+1,family="poisson",x=TRUE)
model1=glm(Y~Z+X,family="poisson")
X=data.frame(X=X)
scr0=compute_gcor(model0 = model0,  X)
scr0$part_cor^2
sc=part_cor_new(model0 = model0,  model1)
sapply(c('v','kl','sse','lr','n'),function(type) rsq.partial(model1,model0,type=type)$partial.rsq)
plot(scr0$IHX,scr0$IHY)

set.seed(1)
Z=rnorm(20)
X=Z+rnorm(20)
Y=rbinom(n=20,1,prob=plogis((-.5*Z+2*X)))
model0=glm(Y~Z+1,family="binomial")
model1=glm(Y~Z+X,family="binomial")
X=data.frame(X=X)
scr0=compute_gcor(model0 = model0,  X)
scr0$part_cor^2
sapply(c('v','kl','sse','lr','n'),function(type) rsq.partial(model1,model0,type=type)$partial.rsq)
plot(scr0$IHX,scr0$IHY)

set.seed(1)
Z=rnorm(20)
X=Z+rnorm(20)
Y=rnorm(n=20,mean=(Z+2*X))
model0=lm(Y~Z+1)
model1=lm(Y~Z+X)
scr0=compute_gcor(model0 = model0,  X)
scr0$part_cor^2
cor(resid(modelX),resid(model0))^2
sapply(c('v','kl','sse','lr','n'),function(type) rsq.partial(model1,model0,type=type)$partial.rsq)
plot(scr0$IHX,scr0$IHY)
library(sensemakr)
partial_r2(model1)



library(pima)
data("pimads")
summary(pimads)
pimads=na.omit(pimads)
mod=glm(diabetes ~ . -npreg,data=pimads,family=binomial,x=TRUE)
scr0=compute_gcor(model0 = mod,  pimads$npreg)
scr0$part_cor^2
plot(scr0$IHX,scr0$IHY)
cor(scr0$IHX,scr0$IHY)^2 
# come mai sono diversi?
# perchè i residui non hanno media 0. perchè?
mean(scr0$IHX)
#[1] -0.0004750917
mean(scr0$IHY)
#[1] -2.90525e-15
# negli lm invece mean(scr0$IHX) è sempre 0!


#esempio 50\% 1 in binomiale. vedi NAGELKERKE 1991
X= Y = rep(0:1,2)
model0=glm(Y~1,family="binomial")
model1=glm(Y~1+X,family="binomial")
X=data.frame(X=X)
scr0=compute_gcor(model0 = model0,  X)
scr0$part_cor^2
sapply(c('v','kl','sse','lr','n'),function(type) rsq.partial(model1,model0,type=type)$partial.rsq)
plot(scr0$IHX,scr0$IHY)
}
\author{
Paolo Girardi and Livio Finos
}
