% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_flipscores.R
\name{confint.flipscores}
\alias{confint.flipscores}
\title{Confidence intervals for flipscores}
\usage{
\method{confint}{flipscores}(
  object,
  parm = NULL,
  level = 0.95,
  type = c("equitailed", "symmetric"),
  score_type = c("standardized", "effective"),
  alternative = c("two.sided", "greater", "less"),
  flips = NULL,
  n_flips = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{flipscores}) model object.}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, they are recovered from the input model (\code{to_be_tested}).}

\item{level}{the confidence level required.}

\item{type}{the type of confidence interval to build, either "equitailed" or "symmetric". Default is "equitailed".}

\item{score_type}{The type of score that is computed. It can be "standardized" or "effective".
Default is inherited by the input model.}

\item{alternative}{The direction of the test to invert, and corresponding interval with respect to the estimate. It can be "two.sided" (default), "greater", or "less".}

\item{flips}{the fixed flips to use in the interval search. If \code{NULL} (default), it tries to inherit it from the input model.}

\item{n_flips}{If \code{flips} are not given or retrieved by the model, the number of random flips of the score contributions. Default inherits them from the input model.}

\item{...}{additional arguments for methods.}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence limits for each parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 in % (by default 2.5% and 97.5%), or "Est." and level or (1-level) if a one-sided alternative is selected.
}
\description{
This function allows to apply the \code{confint} method to \code{flipscores} objects.
}
\details{
All families allowed in \code{glm}, as well as Negative Binomial, are supported, and inherited here by the input model \code{object}. See the main function \code{\link{flipscores}} for further information.

\code{flipscores} is based on randomly sign-flipped score contributions. Performing the test (and generating the confidence interval) different times might give slightly different results. Providing a flip matrix generated with \code{\link{make_flips}} is suggested to uniform the results.

\code{to_be_tested} works differently for factors and character variables: the corresponding dummy variables are to be tested separately. So, to test a numeric variable \code{X}, one could write \code{to_be_tested = "X"}, but to test a factor \code{Z} with levels \code{A}, \code{B} and \code{C} one should write \code{to_be_tested = c("ZB", "ZC")} as they appear in the model summary.

#' @seealso \code{\link{flipscores}}, \code{\link{make_flips}}
}
\examples{
set.seed(1)
dt <- data.frame(X=scale(rnorm(50)),
   Z=factor(rep(LETTERS[1:3],length.out=50)))
dt$Y=rpois(n=nrow(dt),lambda=exp(dt$X*(dt$Z=="C")))
mod0 <- flipscores(Y~Z+X,data=dt,family="poisson")
summary(mod0)
confint(mod0, parm="X")

xx <- rnorm(20)
zz <- rnorm(20, 0.2 * xx)
yy <- rnorm(20, 1 + 2 * xx - zz)
flps <- make_flips(20, 1000)
mod1 <- flipscores(yy ~ xx + zz)
summary(mod1)
confint(mod1, flips = flps)
}
