% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{flipscores-method}
\alias{flipscores-method}
\alias{print.flipscores}
\alias{summary.flipscores}
\title{Methods for flipscores objects}
\usage{
\method{print}{flipscores}(x, ...)

\method{summary}{flipscores}(object, ...)
}
\arguments{
\item{x}{a flipscores object}

\item{...}{additional arguments to be passed}

\item{object}{a flipscores object}
}
\description{
Methods for \code{flipscores} objects.
The following are methods to extract and manipulate relevant information from
a \code{flipscores} object.
}
