% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipscores.R, R/flipscores_engine.R
\name{flipscores}
\alias{flipscores}
\alias{.flipscores_engine}
\title{Flip Scores Test}
\usage{
flipscores(
  formula,
  family = gaussian(),
  data = NULL,
  score_type = "standardized",
  n_flips = 5000,
  alternative = "two.sided",
  id = NULL,
  seed = NULL,
  to_be_tested = NULL,
  tested_coeffs = NULL,
  flips = NULL,
  precompute_flips = TRUE,
  ...
)

flipscores(formula, family, data, score_type = "standardized",
n_flips = 5000, alternative = "two.sided", id = NULL,
seed = NULL, to_be_tested = NULL, flips = NULL,
precompute_flips=TRUE, ...)
}
\arguments{
\item{formula}{see \code{glm} function. It can also be a model (usually generated by a call to \code{glm}); in this case, any other glm-related parameter (e.g. \code{family, data, etc.}) are discarded, the function will make use of the ones used to generate the model.
(i.e. \code{formula}, \code{family}, \code{data}, etc) are not considered.  It is \code{NULL} by default (i.e. not used).}

\item{family}{see \code{glm} function.}

\item{data}{see \code{glm} function.}

\item{score_type}{The type of score that is computed. It can be "standardized" "orthogonalized", "effective" or "basic".
Both "orthogonalized" and "effective" take into account the nuisance estimation and they provide the same
test statistic. In case of small samples "effective score" might have a slight anti-conservative behaviour.
"standardized effective score" gives a solution for this issue.
"orthogonalized" has a similar intent, note however that in case of a big model matrix, it may be slow.}

\item{n_flips}{The number of random flips of the score contributions. Overwritten with the \code{nrow(flips)} when \code{flips} is not \code{NULL} (see parameter \code{flips} for more details).
When \code{n_flips} is equal or larger than the maximum number of possible flips (i.e. n^2), all possible flips are performed.}

\item{alternative}{It can be "greater", "less" or "two.sided" (default)}

\item{id}{a \code{vector} identifying the clustered observations. If \code{NULL} (default) observations are assumed to be independent. If \code{id} is not \code{NULL}, only \code{score_type \%in\% c("standardized","effective","basic")} is allowed, yet.}

\item{seed}{\code{NULL} by default.}

\item{to_be_tested}{vector of indices or names of coefficients of the glm model to be tested (it is faster than computing every scores and p-values of course).}

\item{tested_coeffs}{an alias for  \code{to_be_tested}. \code{tested_coeffs} overcomes \code{to_be_tested} if it is not \code{NULL}.}

\item{flips}{matrix fo +1 or -1, the matrix has \code{n_flips} rows and n (number of observations) columns}

\item{precompute_flips}{\code{TRUE} by default. Overwritten if \code{flips} is not \code{NULL}. If \code{FALSE} the matrix of flips is not computed and the flips are made 'on-the-fly' before computing the test statistics; it may be usefull when \code{flips} is very large (see parameter \code{flips} for more details).
# @param output_flips \code{FALSE} by default. If \code{TRUE} the \code{flips} matrix is returned. Useful when the same flips are needed for more glms, for example in the case of multivariate glms where the joint distribution of test statistis if used for multivariate inference.@param ... see \code{glm} function.}

\item{...}{Additional arguments passed to \code{glm()} or the internal
flipscores engine.}
}
\value{
A \code{flipscores} object (case 1) or a \code{joint_flipscores}
  object (cases 2, 3, 4).

an object of class \code{flipscores}.
See also its methods (\code{summary.flipscores}, \code{anova.flipscores}, \code{print.flipscores}).
}
\description{
Extended version of \code{flipscores()} supporting joint testing when
\code{formula} is a list of glm objects, a list of formulas, or a formula
with a matrix response.

Provides robust tests for testing in GLMs, by sign-flipping score contributions. The tests are often robust against overdispersion, heteroscedasticity and, in some cases, ignored nuisance variables.
}
\details{
\code{flipscores} borrows the same parameters from function \code{glm} (and \code{glm.nb}). See these helps for more details about parameters such as \code{formula},
\code{data}, \code{family}. Note: in order to use Negative Binomial family, \code{family} reference must have quotes (i.e. \code{family="negbinom"}).
 Furthermore, \code{flipscores} object contains two extra elements: \code{scores}  -- i.e. a matrix of n score contributions, one column for each tested coefficient -- and \code{Tspace} -- i.e. a matrix of size \code{n_flips} times \code{ncol(scores)}. The fist row of  \code{Tspace} contains column-wise the test statistics generated by randomly flipping the score contributions, each column refers to the same column of \code{scores}, the vector of observed test statistics (i.e. no flips) is in the first row of \code{Tspace}.
}
\examples{
set.seed(1)
n <- 50
df <- data.frame(
  y1  = rnorm(n),
  y2  = rnorm(n),
  yb  = rbinom(n, 1, 0.5),
  y1p = rpois(n, lambda = 3),
  y2p = rpois(n, lambda = 5),
  x1  = rnorm(n),
  x2  = rnorm(n)
)

# Case 1: single formula
res=flipscores(y1 ~ x1 + x2, family = gaussian(), data = df)
summary(res)

# Case 2: list of glm objects
models <- list(
  glm(y1 ~ x1 + x2, data = df, family = gaussian()),
  glm(y2 ~ x1 + x2, data = df, family = gaussian()),
  glm(yb ~ x1 + x2, data = df, family = binomial())
)
res=flipscores(models)
summary(res)
res=flipscores(models, to_be_tested = "x1")
summary(res)

# Case 3: list of formulas, same family
res=flipscores(
  list(y1 ~ x1 + x2, y2 ~ x1 + x2),
  family = gaussian(),
  data   = df
)
summary(res)

# Case 4: matrix response, gaussian
res=flipscores(cbind(y1, y2) ~ x1 + x2, family = gaussian(), data = df)
summary(res)

# Case 4: matrix response, poisson
res=flipscores(cbind(y1p, y2p) ~ x1 + x2, family = poisson(), data = df)
summary(res)


# more examples
n=20
D=data.frame(X=rnorm(n),Z1=rnorm(n),Z2=rnorm(n))
D$Y=D$Z1+D$X+rnorm(n)
# Run four glms abd combine it in a list
mod1=glm(Y~X+Z1+Z2,data=D)
mod2=glm(Y~X+poly(Z1,2)+Z2,data=D)
mod3=glm(Y~X+poly(Z1,2)+poly(Z2,2),data=D)
mod4=glm(Y~X+Z1+poly(Z2,2),data=D)
mods=list(mod1=mod1,mod2=mod2,mod3=mod3,mod4=mod4)
res=flipscores(mods, to_be_tested = "X")

set.seed(1)
dt=data.frame(X=rnorm(20),
   Z=factor(rep(LETTERS[1:3],length.out=20)))
dt$Y=rpois(n=20,lambda=exp((dt$Z=="C") + 2*dt$X))
mod=flipscores(Y~Z+X,data=dt,family="poisson",n_flips=1000)
summary(mod)

# Equivalent to:
model=glm(Y~Z+X,data=dt,family="poisson")
mod2=flipscores(model)
summary(mod2)

}
\references{
De Santis, Goeman, Hemerik, Davenport, Finos (2025) Inference in generalized linear models with robustness to misspecified variances. JASA <doi:10.1080/01621459.2025.2491775>
}
\seealso{
\code{\link{anova.flipscores}}, \code{\link{summary.flipscores}}, \code{\link[flip]{flip}}, \code{\link{confint.flipscores}}
}
\author{
Livio Finos, Riccardo De Santis, Jesse Hemerik and Jelle Goeman
}
