% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipscores_glm.R
\docType{package}
\name{flipscores_glm}
\alias{flipscores_glm}
\alias{flipscores_glm-package}
\title{Performing customized testing with GLM's estimation}
\usage{
flipscores_glm(formula, family, data, score_type = "orthogonalized", n_flips=1000, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{family}{a description of the error distribution and link function to be used in the model.
For glm this can be a character string naming a family function, a family function or the result of a call to a family function.
For glm.fit only the third option is supported. OK?--> Note: to use Negative Binomial family, family string reference must have quotes.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame)
containing the variables in the model. If not found in data, the variables are taken from environment(formula),
typically the environment from which glm is called.}

\item{score_type}{The type of score that is computed, either "basic", "effective" or "orthogonalized". Using "effective" takes into account nuisance estimation. ORTHO?? Default is "orthogonalized".}

\item{alternative}{Should be "greater", "less" or "two.sided". By default is "two.sided"}

\item{statTest}{C'e ANCORA?? Choose a test statistic from flip.statTest. See "flip" package.}

\item{testType}{C'e ANCORA?? by default testType="permutation". The use of option "combination" is more efficient when
X is indicator of groups (i.e. C>1 samples testing). When the total number of possible combinations exceeds 10 thousand,
"permutation" is performed. As an alternative, if you choose "rotation", resampling is performed through random linear
combinations (i.e. a rotation test is performed). This option is useful when only few permutations are available, that is,
minimum reachable significance is hight. See also the details section for the algorithm used. The old syntax rotationTest=TRUE
is maintained for compatibility but is deprecated, use testType="rotation" instead.}

\item{n_perms}{The number of times that the scores are randomly sign-flipped. Note: the maximum number of possible permutation is n!^p.
Where n indicates the number of observations (rows) and p indicates the number of
covariates (columns). R typing: factorial(n)^p. Default is 1000..}
}
\value{
flipscores_glm returns an object of class inheriting from "glm" which inherits from the class "lm". See also glm package.  
The functions related (summary.flipscores, plot.flipscores, print.flipscores) can be used to obtain or print a summary of the results, whose include customized p.values.
}
\description{
Used to both fit generalized linear models and perform many tests on coefficients.
All of them based on flipped likelihood scores. Referring to canonical glm function options,
this one integrates that with more coefficient testing ways: test statistic type, score type,
resampling type and test direction. The resulting output is same as the late,
but p-value is calculated accounting to new preferencies.
}
\examples{
data(iris)
data=iris[iris$Species!="setosa",]
data$Species=factor(data$Species)
m1 = flipscores_glm(Species~.+Petal.Width*Petal.Length, family=binomial(link = "logit"), data=data, score_type="orthogonalized", n_flips=1000)
summary(m1)

data = as.data.frame(Titanic)
m1 = flipscores_glm(Freq~., family=poisson(link = "log"), data=data, score_type="orthogonalized", n_flips=1000)
summary(m1)

}
\seealso{
flip
}
\author{
Jesse Hemerik, Jelle Goeman and Livio Finos, with contribution by Vittorio Giatti and Angela Andreella
}
