% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_head_flip_out}
\alias{get_head_flip_out}
\title{plot.flipscores <- function (x, ...) {
  summary.flipscores(x, ...)
  # object <- x
  # dots <- list(...)
  # a <- object$plotting
  # b <- object$Descriptive
  # fac.names <- a$fac_names
  # exist <- hasArg(factor)
  #
  # if(length(fac.names) != 1){
  #   if(!exist){
  #     print("Please choose the factor you wish to plot (for interaction type something like group1:group2) and confirm by pressing 'Enter'")
  #     Faktor <- scan("", what="character")
  #     while(length(Faktor)==0){
  #       print("Please enter the name of the factor you wish to plot!")
  #       Faktor <- scan("", what="character")
  #     }
  #     } else {
  #      Faktor <- dots$factor
  #     }
  #   x.label <- ""
  # } else {
  #   Faktor <- fac.names
  #   x.label <- fac.names
  # }
  #
  # match.arg(Faktor, fac.names)
  #
  # # default values
  # args <- list(plot.object = a, descr.object = b, factor = Faktor,
  #              lwd =2, ylab = "Means", xlab = x.label, col = 1:length(fac.names), pch = 1:18, legendpos = "topright")
  #
  # args[names(dots)] <- dots
  #
  # do.call(plotting, args = args)
}}
\usage{
get_head_flip_out(x)
}
\description{
plot.flipscores <- function (x, ...) {
  summary.flipscores(x, ...)
  # object <- x
  # dots <- list(...)
  # a <- object$plotting
  # b <- object$Descriptive
  # fac.names <- a$fac_names
  # exist <- hasArg(factor)
  #
  # if(length(fac.names) != 1){
  #   if(!exist){
  #     print("Please choose the factor you wish to plot (for interaction type something like group1:group2) and confirm by pressing 'Enter'")
  #     Faktor <- scan("", what="character")
  #     while(length(Faktor)==0){
  #       print("Please enter the name of the factor you wish to plot!")
  #       Faktor <- scan("", what="character")
  #     }
  #     } else {
  #      Faktor <- dots$factor
  #     }
  #   x.label <- ""
  # } else {
  #   Faktor <- fac.names
  #   x.label <- fac.names
  # }
  #
  # match.arg(Faktor, fac.names)
  #
  # # default values
  # args <- list(plot.object = a, descr.object = b, factor = Faktor,
  #              lwd =2, ylab = "Means", xlab = x.label, col = 1:length(fac.names), pch = 1:18, legendpos = "topright")
  #
  # args[names(dots)] <- dots
  #
  # do.call(plotting, args = args)
}
}
