% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_flipscores_methods.R, R/methods.R
\docType{methods}
\name{jfs-methods}
\alias{jfs-methods}
\alias{p.adjust}
\title{Methods for jfs objects}
\usage{
p.adjust(object, method = "maxT", tail = 0, ...)
}
\arguments{
\item{object}{an object of class \code{jfs} and \code{flipscores}.}

\item{method}{any method implemented in \code{flip.adjust} or
a custom function. In the last case it must be a function that uses a matrix
as input and returns a vector of adjusted p-values equal to the number of columns of the inputed matrix.}

\item{tail}{argument: expresses the tail direction of the alternative hypothesis.
It can be "two.sided" (or 0, the default), "less" (or -1) or "greater" (or +1).}

\item{...}{additional arguments to be passed}
}
\description{
Collection of methods for objects of class \code{jfs}.

\code{p.adjust} method for class "\code{jfs}" and "\code{flipscores}".
Add adjusted p-values into the \code{jfs}and \code{flipscores} object.
}
