% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_flipscores_methods.R
\name{plot.joint_flipscores}
\alias{plot.joint_flipscores}
\title{Plot method for joint_flipscores}
\usage{
\method{plot}{joint_flipscores}(x, ...)
}
\arguments{
\item{x}{A \code{joint_flipscores} object.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\description{
Plots the score distributions for each model.
}
