% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_flipscores_methods.R
\name{print.jfs}
\alias{print.jfs}
\title{Print method for jfs}
\usage{
\method{print}{jfs}(x, n = 2, ...)
}
\arguments{
\item{x}{A \code{jfs} object.}

\item{n}{Number of rows to show at head and tail of summary table.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for jfs
}
