% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_flipscores_methods.R
\name{print.joint_flipscores}
\alias{print.joint_flipscores}
\title{Print method for joint_flipscores}
\usage{
\method{print}{joint_flipscores}(x, ...)
}
\arguments{
\item{x}{A \code{joint_flipscores} object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for joint_flipscores
}
