% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summary.flipscores}
\alias{summary.flipscores}
\title{summary.flipscores summary method for a flipscores-object.}
\usage{
summary.flipscores(object, ...)
}
\arguments{
\item{object}{a flipscores-object}

\item{...}{additional arguments to be passed}
}
\description{
summary.flipscores summary method for a flipscores-object.
}
