% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_flipscores_methods.R
\name{summary.jfs}
\alias{summary.jfs}
\title{Summary method for jfs}
\usage{
\method{summary}{jfs}(object, digits = 4, ...)
}
\arguments{
\item{object}{A \code{jfs} object.}

\item{digits}{Number of digits to print. Default \code{4}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Summary method for jfs
}
