% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_flipscores_methods.R
\name{summary.joint_flipscores}
\alias{summary.joint_flipscores}
\title{Summary method for joint_flipscores}
\usage{
\method{summary}{joint_flipscores}(object, ...)
}
\arguments{
\item{object}{A \code{joint_flipscores} object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Summary method for joint_flipscores
}
