% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores.R
\name{compute_scores}
\alias{compute_scores}
\title{compute_scores}
\usage{
compute_scores(fit0, X, score_type = "orthogonalized")
}
\description{
compute_scores
}
\examples{

Z=rnorm(20)
X=Z+rnorm(20)
Y=rpois(n=20,lambda=exp(Z+X))
mod0=glm(Y~Z,family="poisson")
scr0=compute_scores(fit0 = mod0, X = X, score_type = "basic")
flip:::flip(scr0)

mod1=glm(Y~Z+X,family="poisson")
scr1=compute_scores(fit0 = mod1, X = X, score_type = "basic")
flip:::flip(scr1)

}
