% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipscores.R
\docType{package}
\name{flipscores}
\alias{flipscores}
\alias{flipscores-package}
\title{Robust testing in GLMs}
\usage{
flipscores(model0, model1, X1, alternative = "two.sided",  w=1E5, scoretype="basic")
}
\arguments{
\item{score_type}{The type of score that is computed, either "orthogonalized", "effective" or "basic". 
By default is "orthogonalized". 
"effective" and "orthogonalized" takes into account nuisance estimation.}

\item{n_flips}{The number of random flips of the scores contributions
When \code{n_flips} is equal or larger than the maximum number of possible flips (i.e. n^2), all possible flips are performed. 
Default is 5000.}

\item{id}{a \code{vector} identifying the clustered observations. If \code{NULL} (default) observations are assumed to be independent.}

\item{alternative}{Should be "greater", "less" or "two.sided". DEFAULT IS??}

\item{alternative}{Should be "greater", "less" or "two.sided". By default is "two.sided"}
}
\value{
glm class object with sign-flip score test.
See also the related functions (summary.flipscores, plot.flipscores, print.flipscores).
}
\description{
Provides two robust tests for testing in GLMs, by sign-flipping score contributions. The tests are often robust against overdispersion, heteroscedasticity and, in some cases, ignored nuisance variables.

Used to both fit generalized linear models and perform a sign-flip score test on coefficients.
}
\details{
\code{flipscores} borrow the same parameters from function \code{glm} (and \code{glm.nb}). See these helps for more datail about parameters such as \code{formula}
\code{data},\code{family}. Note: in order to use Negative Binomial family, \code{family} reference must have quotes (i.e. \code{family="negbinom"}).
}
\examples{
data(iris)
data=iris[iris$Species!="setosa",]
data$Species=factor(data$Species)
m1 = flipscores(Species~.+Petal.Width*Petal.Length, family=binomial(link = "logit"), data=data, score_type="orthogonalized", n_flips=1000)
summary(m1)

data = as.data.frame(Titanic)
m1 = flipscores(Freq~., family=poisson(link = "log"), data=data, score_type="orthogonalized", n_flips=1000)
summary(m1)

}
\references{
"Robust testing in generalized linear models by sign-flipping score contributions" by J.Hemerik, J.Goeman and L.Finos.
}
\seealso{
flip
}
\author{
Livio Finos and Vittorio Giatti
}
