% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipscores_glm.R
\docType{package}
\name{flipscores_glm}
\alias{flipscores_glm}
\alias{flipscores_glm-package}
\title{Performing robust testing with GLM's estimation}
\usage{
flipscores_glm(formula, family, data, score_type = "orthogonalized", n_flips=1000, ...)
}
\arguments{
\item{formula}{an object of class "formula" of the model to be fitted.}

\item{family}{distribution of the response variable. Note: in order to use Negative Binomial family, family reference must have quotes (family="negbinom").}

\item{data}{data frame used. A list or environment (or object coercible by as.data.frame to a data frame)
containing the variables in the model.}

\item{score_type}{The type of score that is computed, either "basic", "effective" or "orthogonalized". By default is "orthogonalized". 
#' Using "effective" and "orthogonalized" takes into account nuisance estimation.}

\item{n_flips}{The number of random flips of the scores contributions
When \code{n_flips} is equal or larger than the maximum number of possible flips (i.e. n^2), all possible flips are performed. 
Default is 5000.}

\item{cluster}{identifies subgrouped observations. Note: must be a vector.}

\item{alternative}{Should be "greater", "less" or "two.sided". By default is "two.sided"}
}
\value{
glm class object with sign-flip score test.
See also the related functions (summary.flipscores, plot.flipscores, print.flipscores).
}
\description{
Used to both fit generalized linear models and perform a sign-flip score test on coefficients.
}
\examples{
data(iris)
data=iris[iris$Species!="setosa",]
data$Species=factor(data$Species)
m1 = flipscores_glm(Species~.+Petal.Width*Petal.Length, family=binomial(link = "logit"), data=data, score_type="orthogonalized", n_flips=1000)
summary(m1)

data = as.data.frame(Titanic)
m1 = flipscores_glm(Freq~., family=poisson(link = "log"), data=data, score_type="orthogonalized", n_flips=1000)
summary(m1)

}
\references{
"Robust testing in generalized linear models by sign-flipping score contributions" by J.Hemerik, J.Goeman and L.Finos.
}
\seealso{
flip
}
\author{
Livio Finos and Vittorio Giatti
}
