% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flips.R
\name{make_flips}
\alias{make_flips}
\title{It creates a \code{n_flips}x\code{n_obs} matrix of random +1 and -1. 
The first row is made by ones (i.e. the observed test statistic is computed)}
\usage{
make_flips(n_obs, n_flips)
}
\arguments{
\item{n_obs}{number of observations}

\item{n_flips}{number of flips}
}
\description{
It creates a \code{n_flips}x\code{n_obs} matrix of random +1 and -1. 
The first row is made by ones (i.e. the observed test statistic is computed)
}
