% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_flipscores_methods.R, R/utils_jointest.R
\name{print.joint_flipscores}
\alias{print.joint_flipscores}
\title{Print method for joint_flipscores}
\usage{
\method{print}{joint_flipscores}(x, n = 2, ...)

\method{print}{joint_flipscores}(x, n = 2, ...)
}
\arguments{
\item{x}{A \code{joint_flipscores} object.}

\item{n}{Number of rows to show at head and tail of summary table.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for joint_flipscores

Print method for joint_flipscores
}
