% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_flipscores_methods.R, R/utils_jointest.R
\name{summary.joint_flipscores}
\alias{summary.joint_flipscores}
\title{Summary method for joint_flipscores}
\usage{
\method{summary}{joint_flipscores}(object, digits = 4, ...)

\method{summary}{joint_flipscores}(object, digits = 4, ...)
}
\arguments{
\item{object}{A \code{joint_flipscores} object.}

\item{digits}{Number of digits to print. Default \code{4}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Summary method for joint_flipscores

Summary method for joint_flipscores
}
